Converts graphql schema and django backend to asychronous (asgi)
