+++ dependencies.txt
@@ -3,26 +3,41 @@ ariadne==0.14.1
 ariadne-django==0.2.0
 ariadne-jwt==0.1.7
 asgiref==3.4.1
+attrs==21.4.0
+autobahn==21.2.1
+Automat==20.2.0
+cffi==1.15.0
+channels==3.0.4
+constantly==15.1.0
 contextlib2==21.6.0
 contextvars==2.4
+cryptography==36.0.2
+daphne==3.0.2
 dataclasses==0.8
 Django==3.2.12
 django-cors-headers==3.10.1
 graphql-core==3.1.7
+hyperlink==21.0.0
 idna==3.3
 immutables==0.16
-pkg-resources==0.0.0
+incremental==21.3.0
+pkg_resources==0.0.0
 promise==2.3
+pyasn1==0.4.8
+pyasn1-modules==0.2.8
+pycparser==2.21
 PyJWT==2.3.0
+pyOpenSSL==22.0.0
 python-dateutil==2.8.2
 pytz==2022.1
 pyzmq==22.3.0
+service-identity==21.1.0
 six==1.16.0
 sniffio==1.2.0
 sqlparse==0.4.2
 starlette==0.17.1
-typing-extensions==4.1.1
+Twisted==22.4.0
+txaio==22.2.1
+typing_extensions==4.1.1
 zmq==0.0.0
-LabJackPython==2.0.4
-numpy==1.19.5
-pyserial==3.5
+zope.interface==5.4.0
