+++ nEDM_server/nEDM_server/urls.py
@@ -16,10 +16,6 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import path
 
-from ariadne_django.views import GraphQLView
-from .graphql_config import schema
-
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('graphql/', GraphQLView.as_view(schema=schema)),
 ]
